+++ nixos/default.nix
@@ -1,8 +1,8 @@
 {
   imports = [
-    ./modules/environment.nix
     ./modules/nix.nix
     ./modules/ranger.nix
+    ./modules/shell.nix
     ./modules/tmux.nix
     ./modules/utilities.nix
     ./modules/vim.nix